tools/benchmark/requirements.txt
onnxruntime
tensorrt
pycuda
tqdm
